+++ src/connection/publisher.js
@@ -1,13 +1,12 @@
 /* @flow */
 import ConnectionBase from './base';
 
-class ConnectionPublisher extends ConnectionBase {
+export default class ConnectionPublisher extends ConnectionBase {
   connect(stream: ?MediaStream.prototype) {
     this.role = 'upstream';
     if (this.options && this.options.multistream) {
       return this._multiStream(stream);
-    }
-    else {
+    } else {
       return this._singleStream(stream);
     }
   }
@@ -23,8 +22,7 @@ class ConnectionPublisher extends ConnectionBase {
           stream.getTracks().forEach(track => {
             this._pc.addTrack(track, stream);
           });
-        }
-        else {
+        } else {
           this._pc.addStream(stream);
         }
         this.stream = stream;
@@ -44,14 +42,6 @@ class ConnectionPublisher extends ConnectionBase {
       .then(this._signaling.bind(this))
       .then(this._connectPeerConnection.bind(this))
       .then(message => {
-        if (typeof this._pc.addStream === 'undefined') {
-          stream.getTracks().forEach(track => {
-            this._pc.addTrack(track, stream);
-          });
-        }
-        else {
-          this._pc.addStream(stream);
-        }
         if (typeof this._pc.ontrack === 'undefined') {
           this._pc.onaddstream = event => {
             if (this.connectionId !== event.stream.id) {
@@ -78,8 +68,18 @@ class ConnectionPublisher extends ConnectionBase {
           }
           this._callbacks.removestream(event);
         };
-        this.stream = stream;
         return this._setRemoteDescription(message);
+      })
+      .then(message => {
+        if (typeof this._pc.addStream === 'undefined') {
+          stream.getTracks().forEach(track => {
+            this._pc.addTrack(track, stream);
+          });
+        } else {
+          this._pc.addStream(stream);
+        }
+        this.stream = stream;
+        return Promise.resolve(message);
       })
       .then(this._createAnswer.bind(this))
       .then(this._sendAnswer.bind(this))
@@ -89,5 +89,3 @@ class ConnectionPublisher extends ConnectionBase {
       });
   }
 }
-
-module.exports = ConnectionPublisher;