+++ src/sora.js
@@ -1,35 +1,31 @@
 /* @flow */
 import ConnectionPublisher from './connection/publisher';
 import ConnectionSubscriber from './connection/subscriber';
-import type { ConnectionOptions } from './connection/base';
+import type { ConnectionOptions } from './utils';
 
-
-const Sora = {
-  connection: function(signalingUrl: string, debug: boolean=false) {
+export default {
+  connection: function(signalingUrl: string, debug: boolean = false) {
     return new SoraConnection(signalingUrl, debug);
   },
   version: function() {
-    return process.version;
+    return SORA_JS_SDK_VERSION;
   }
 };
-
 
 class SoraConnection {
   signalingUrl: string;
   debug: boolean;
 
-  constructor(signalingUrl: string, debug: boolean=false) {
+  constructor(signalingUrl: string, debug: boolean = false) {
     this.signalingUrl = signalingUrl;
     this.debug = debug;
   }
 
-  publisher(channelId: string, metadata: string, options: ConnectionOptions={ audio: true, video: true }) {
+  publisher(channelId: string, metadata: string, options: ConnectionOptions = { audio: true, video: true }) {
     return new ConnectionPublisher(this.signalingUrl, channelId, metadata, options, this.debug);
   }
 
-  subscriber(channelId: string, metadata: string, options: ConnectionOptions={ audio: true, video: true }) {
+  subscriber(channelId: string, metadata: string, options: ConnectionOptions = { audio: true, video: true }) {
     return new ConnectionSubscriber(this.signalingUrl, channelId, metadata, options, this.debug);
   }
 }
-
-module.exports = Sora;